backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pgvector
groq
pydantic
python-dotenv